% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{all_ambiguous_sequences}
\alias{all_ambiguous_sequences}
\title{If a set of words is not a code it returns all ambiguous sequences.}
\usage{
all_ambiguous_sequences(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
A String vector with all ambiguous sequences.
}
\description{
This function returns all ambiguous sequences
which only exist if a set of words \emph{X} is by
definition not a code. Such a sequence can be decomposed in
at least two disjoint sets of words of \emph{X}.
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
all_ambiguous_sequences(code)

}
\seealso{
\link{is_code}
}
