% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_c3_codes}
\alias{all_c3_codes}
\title{List of all 216 maximal self complementary circular C3 codes.}
\format{
A list with 216 entries:
\describe{
\item{code}{A maximal self complementary circular code}
}
}
\usage{
all_c3_codes
}
\description{
List of all 216 maximal self complementary circular C3 codes.
}
\keyword{datasets}
