% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{c3_code}
\alias{c3_code}
\title{The i-th C3 code.}
\usage{
c3_code(i)
}
\arguments{
\item{i}{Integer 1 <= i <= 216. The number of the C3 code}
}
\value{
A C3 code
}
\description{
Returns the i-th maximal self-complmentary C3 code. There are exactly 216 such codes.
}
