% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{c3_codes}
\alias{c3_codes}
\title{The i-th C3 code.}
\usage{
c3_codes(is = seq_along(all_c3_codes))
}
\arguments{
\item{is}{a vector of Integer. for all i in is: 1 <= i <= 216.}
}
\value{
A list of C3 codes
}
\description{
Returns a list of maximal self-complementary C3 codes. Without paramet it returns all C3 codes.
}
