% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\docType{data}
\name{c3_equivmatrix}
\alias{c3_equivmatrix}
\title{Table for mapping of code numbers to equivalence classes.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 216 rows and 2 columns.
}
\usage{
c3_equivmatrix
}
\value{
First column: C3 code number, second row: equivalence class.
}
\description{
Table for mapping of code numbers to equivalence classes.
}
\keyword{datasets}
