% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{c3_in_class}
\alias{c3_in_class}
\title{All code numbers for given equivalence class}
\usage{
c3_in_class(eid)
}
\arguments{
\item{eid}{Equivalence class.}
}
\value{
List of C3 code numbers.
}
\description{
All code numbers for given equivalence class
}
