% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{circular_shift}
\alias{circular_shift}
\title{Shifts each tuple by \code{sh} positions}
\usage{
circular_shift(tuples, sh)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}

\item{sh}{A integer, the shift index, i.e. the number of shifts.}
}
\value{
Boolean value. True if the code is circular.
}
\description{
Under the concept shift is understood a circular permutation, i.e.
let \emph{X}={123, 332}, then c.shift(2) results in {312, 233}
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
circular_shift(code, 2)

}
