% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_cyclic_paths}
\alias{get_cyclic_paths}
\title{Returns a list of all cyclic paths}
\usage{
get_cyclic_paths(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
A list of String vectors with all cyclic paths
}
\description{
This function returns all cyclic paths
in the graph associated to a set of words \emph{X}.
}
\examples{
code <- gcatbase::code(c("ACG", "CGA", "CA"))
lp <- get_cyclic_paths(code)

}
\seealso{
\link{get_representing_graph}
}
