% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_exact_k_circular}
\alias{get_exact_k_circular}
\title{This function checks if a code is k-circular.}
\usage{
get_exact_k_circular(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
Integer value, the exact k value of the k-circularity.
}
\description{
K circle codes are a less restrictive code from the family of circle codes. These codes only ensure that for every
concatenation of less than k tuples from \emph{X} written on a circle, there is only one partition in tuples from \emph{X}.\cr
For mor details see: \link{https://link.springer.com/article/10.1007/s11538-020-00770-7}
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
k <- get_exact_k_circular(code)

}
\seealso{
\link{is_code_circular}
}
