% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_k_graph_circular}
\alias{get_k_graph_circular}
\title{This function checks if a code is K-Graph circular.}
\usage{
get_k_graph_circular(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
Integer value, the k-graph value of the k-graph-circularity. If code is not k-graph circle it returns -1.
}
\description{
K-graph circle codes are a more restrictive than k-circle codes.
These codes only ensure that all cycles in the representing graph have a common length.
If the code is circular or the code contains cycles of
different length the teh function returns -1. \cr
For mor details see: \link{https://link.springer.com/article/10.1007/s11538-020-00770-7}
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
k <- get_k_graph_circular(code)

}
\seealso{
\link{get_exact_k_circular}
}
