% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_longest_paths}
\alias{get_longest_paths}
\title{Returns a list of all longest paths}
\usage{
get_longest_paths(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
A list of String vectors with all longest paths.
}
\description{
This function returns all longest paths
in the graph associated to a set of words \emph{X}.
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
lp <- get_longest_paths(code)

}
\seealso{
\link{get_representing_graph}
}
