% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_representing_component_obj}
\alias{get_representing_component_obj}
\title{Returns a i-component associated to a code}
\usage{
get_representing_component_obj(tuples, i, show_cycles, show_longest_path)
}
\arguments{
\item{tuples}{a gcatbase::gcat.code object}

\item{i}{a integer the component index}

\item{show_cycles}{a boolean, if true all edges in all cyclic paths a red}

\item{show_longest_path}{a boolean, if true all edges in all longest paths a red}
}
\value{
a i-component rust graph-object associated to a code
}
\description{
Returns a i-component associated to a code
}
\examples{
code <- gcatbase::code(c("ACGC", "CGGG", "AC"))
g <- get_representing_component_obj(code,2)

}
