% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.rust.wrapper.R
\name{get_representing_graph}
\alias{get_representing_graph}
\title{Prepares a R igraph object of a graph associated to a code.}
\usage{
get_representing_graph(code, show_cycles = F, show_longest_path = F)
}
\arguments{
\item{code}{is A gcatbase::gcat.code object.}

\item{show_cycles}{A bool value. If true the all edges which are part of a cycle are colored red.}

\item{show_longest_path}{A bool value. If true the all edges part of the longest path are colored blue.}
}
\value{
A igraph (\url{http://igraph.org/r/}) object: A graph representing a circular code.
}
\description{
This function factors a igraph (\url{http://igraph.org/r/}) object of an representing graph of a circular code.
The following definition describes a directed graph to an  code.
Recall from graph theory (Clark and Holton, 1991) that a graph G consists of
a finite set of vertices (nodes) V and a finite set of edges E. Here, an edge is a set \{v,w\} of vertices
from V . The graph is called oriented if the edges have an orientation, i.e. edges are considered to be
ordered pairs \link{v,w} in this case.\cr
Definition Let X be a code. We define a directed graph G(X) =
(V (X),E(X)) with set of vertices V (X) and set of edges E(X) as follows:
V (X) = \{N1...Ni,Ni+1...Nn : N1N2N3...Nn in X, 0 < i < n\}\cr
E(X) = \{\link{N1...Ni,Ni+1...Nn} : N1N2N3...Nn in X, 0 < i < n\}\cr
The graph G(X) is called the representing graph of X or the graph associated to X.\cr
Basically, the graph G(X) associated to a code X interprets n-tuples from X in (n−1) ways
by pairs of i-tuples and (n-i)-tuples for 0 < i < n.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
G <- get_representing_graph(code, TRUE, TRUE)
igraph::tkplot(G)

}
