% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_representing_graph_obj}
\alias{get_representing_graph_obj}
\title{Returns the graph associated to a code}
\usage{
get_representing_graph_obj(tuples, show_cycles, show_longest_path)
}
\arguments{
\item{tuples}{a gcatbase::gcat.code object}

\item{show_cycles}{a boolean, if true all edges in all cyclic paths a red}

\item{show_longest_path}{a boolean, if true all edges in all longest paths a red}
}
\value{
a rust graph-object associated to a code
}
\description{
Returns the graph associated to a code
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
g <- get_representing_graph_obj(code,2)

}
