% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{is_code_circular}
\alias{is_code_circular}
\title{Check if a code is circular.}
\usage{
is_code_circular(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
Boolean value. True if the code is circular.
}
\description{
This function checks if a code is circular. Circular codes are sets of
tuples \emph{X} of different tuple length where
every concatenation of words \emph{w} in \emph{X} written on a circle
has only a single decomposition into words from \emph{X}.\cr
For more info on this subject read:\cr
\link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5492142/},\cr
\link{http://dpt-info.u-strasbg.fr/~c.michel/Circular_Codes.pdf},\cr
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
is_code_circular(code)

}
