% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{is_code_cn_circular}
\alias{is_code_cn_circular}
\title{This function checks if a code is Cn-circular.}
\usage{
is_code_cn_circular(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
Boolean value. True if the code is Cn circular.
}
\description{
That all circular permutations of the code (of all tuples) are circular codes again.
In total, this function checks all 'n' circular permutations where 'n' is the greatest
common multiple of all tuple lengths used.
This is an extended property of circular codes.
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
k <- is_code_cn_circular(code)

}
\seealso{
\link{is_code_circular}
}
