% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{is_code_strong_comma_free}
\alias{is_code_strong_comma_free}
\title{Check if a code is strong comma free.}
\usage{
is_code_strong_comma_free(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
Boolean value. True if the code is strong comma free.
}
\description{
This function checks if a code is strong comma free.
Strong comma free codes are a more restrictive codes from the circular code family.
A strong comma free code \emph{X} is a code in which no nonempty suffix of any word from \emph{X}
is a nonempty prefix of any word from \emph{X}.\cr
This is an extended property of the circular codes. See \link{is_code_comma_free} for more details.\cr
For more info on this subject read:\cr
\link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5492142/},\cr
\link{http://dpt-info.u-strasbg.fr/~c.michel/Circular_Codes.pdf},\cr
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
is_code_strong_comma_free(code)

}
\seealso{
\link{is_code_circular}, \link{is_code_comma_free}
}
