% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{c3_equiv_class}
\alias{c3_equiv_class}
\title{Equivalence class for a C3 code number.}
\usage{
c3_equiv_class(cid)
}
\arguments{
\item{cid}{Integer 0 < i < 217. The number of the C3 code.}
}
\value{
Its equivalence class.
}
\description{
Equivalence class for a C3 code number.
}
