% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{is_code_cn_circular}
\alias{is_code_cn_circular}
\title{This function checks if a code is Cn-circular.}
\usage{
is_code_cn_circular(tuples)
}
\arguments{
\item{tuples}{A gcatbase::gcat.code object}
}
\value{
Boolean value. True if the code is Cn circular.
}
\description{
A code is cn circular if all circular permutations of the code (of all tuples) are circular codes again.
In total, this function checks 'x' circular permutations where 'x' is the least
common multiple of all tuple lengths used. This is an extended property of circular codes.
}
\examples{
code <- gcatbase::code(c("ACG", "CGG", "AC"))
k <- is_code_cn_circular(code)

}
\seealso{
\link{is_code_circular}
}
